move known values around
