+++ src/app/inventory/spreadsheets.ts
@@ -19,6 +19,7 @@ import { getRating } from '../item-review/reducer';
 import { getSpecialtySocketMetadata, getMasterworkStatNames } from 'app/utils/item-utils';
 import store from '../store/store';
 import { t } from 'app/i18next-t';
+import { dimArmorStatHashByName } from 'app/search/search-filter-values';
 
 // step node names we'll hide, we'll leave "* Chroma" for now though, since we don't otherwise indicate Chroma
 const FILTER_NODE_NAMES = [
@@ -285,16 +286,6 @@ export function source(item: DimItem) {
   }
 }
 
-export const armorStatHashes = {
-  Mobility: 2996146975,
-  Resilience: 392767087,
-  Recovery: 1943323491,
-  Discipline: 1735777505,
-  Intellect: 144602215,
-  Strength: 4244567218,
-  Total: -1000,
-};
-
 function downloadArmor(items: DimItem[], nameMap: { [key: string]: string }, itemInfos: ItemInfos) {
   // We need to always emit enough columns for all perks
   const maxPerks = getMaxPerks(items);
@@ -384,9 +375,9 @@ function downloadArmor(items: DimItem[], nameMap: { [key: string]: string }, ite
       row.Disc = stats.Discipline ? stats.Discipline.value : 0;
       row.Str = stats.Strength ? stats.Strength.value : 0;
     } else {
-      const armorStats = Object.keys(armorStatHashes).map((statName) => ({
+      const armorStats = Object.keys(dimArmorStatHashByName).map((statName) => ({
         name: statName,
-        stat: stats[armorStatHashes[statName]],
+        stat: stats[dimArmorStatHashByName[statName]],
       }));
       armorStats.forEach((stat) => {
         row[stat.name] = stat.stat ? stat.stat.value : 0;
