+++ src/app/inventory/store/stats.ts
@@ -16,7 +16,7 @@ import { compareBy } from 'app/utils/comparators';
 import _ from 'lodash';
 import { t } from 'app/i18next-t';
 import { getSocketsWithStyle, getSocketsWithPlugCategoryHash } from '../../utils/socket-utils';
-import { D2CategoryHashes } from 'app/search/search-filter-values';
+import { D2ItemCategoryHashes } from 'app/search/d2-known-values';
 
 /**
  * These are the utilities that deal with Stats on items - specifically, how to calculate them.
@@ -167,7 +167,7 @@ function buildStatsFromMods(
 ): DimStat[] {
   const statTracker: { stat: number; value: number } | {} = {};
   const investmentStats: DimStat[] = [];
-  const modSockets = getSocketsWithPlugCategoryHash(itemSockets, D2CategoryHashes.armormod);
+  const modSockets = getSocketsWithPlugCategoryHash(itemSockets, D2ItemCategoryHashes.armormod);
   const masterworkSockets = getSocketsWithStyle(
     itemSockets,
     DestinySocketCategoryStyle.EnergyMeter
